backend/requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
apturl==0.5.2
asgiref==3.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
backcall==0.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
bleach==3.1.1
blinker==1.4
Brlapi==0.7.0
cachetools==5.2.0
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
cloud-init==22.2
cloudinary==1.44.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dataframe-image==0.1.2
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==3.0.1
Django==4.2.27
django-cors-headers==4.4.0
djangorestframework==3.15.2
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.8.0
flatbuffers==1.12
fonttools==4.37.1
frozenlist==1.3.1
future==0.18.2
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==4.0.0rc1
grpcio==1.48.1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
h5py==3.7.0
hpack==3.0.0
hstspreload==2022.9.1
html5lib==1.0.1
httpcore==0.9.1
httplib2==0.14.0
httpx==0.13.3
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.12.0
incremental==16.10.1
ipykernel==5.2.0
ipython==7.13.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
jedi==0.15.2
Jinja2==2.10.1
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.0.0
jupyter-core==4.6.3
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.6
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
multidict==6.0.2
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
notebook==6.0.3
numpy==1.23.2
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==2.5.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.19.4
psycopg2-binary==2.9.10
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
PyPDF2==2.10.6
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==1.0.1
pytz==2022.2.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==5.3.2
QtPy==2.2.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
SecretStorage==2.3.1
selenium==4.4.3
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
sos==4.3
soupsieve==2.3.2.post1
sqlparse==0.5.5
ssh-import-id==5.10
systemd-python==234
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.2
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
threadpoolctl==3.1.0
tornado==5.1.1
traitlets==4.3.3
trio==0.21.0
trio-websocket==0.9.2
Twisted==18.9.0
typing_extensions==4.3.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.7
virtualenv==20.16.4
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.2.2
whitenoise==6.7.0
widgetsnbextension==2.0.0
wrapt==1.14.1
wsproto==1.2.0
xkit==0.0.0
xlrd==2.0.1
yarl==1.8.1
youtube-dl==2021.12.17
zipp==1.0.0
zope.interface==4.7.1
